Simplify monitoring to single session
