+++ app/modules/AgentHubCore/Tests/AgentHubTests/MonitoringItemsSnapshotTests.swift
@@ -5,7 +5,6 @@ import Testing
 
 private struct SnapshotItem: Identifiable {
   let id: String
-  let modulePath: String
   let timestamp: Date
 }
 
@@ -15,46 +14,37 @@ struct MonitoringItemsSnapshotTests {
   func singleLayoutUsesPrimaryItem() throws {
     let older = SnapshotItem(
       id: "claude-old",
-      modulePath: "/tmp/repo-a",
       timestamp: Date(timeIntervalSince1970: 100)
     )
     let newer = SnapshotItem(
       id: "codex-new",
-      modulePath: "/tmp/repo-b",
       timestamp: Date(timeIntervalSince1970: 200)
     )
 
     let snapshot = MonitoringItemsSnapshot(
       items: [older, newer],
       primaryItemID: older.id,
-      layoutMode: .single,
-      modulePath: { $0.modulePath },
       timestamp: { $0.timestamp }
     )
 
     #expect(snapshot.effectivePrimaryItemID == older.id)
     #expect(snapshot.visibleItems.map(\.id) == [older.id])
-    #expect(snapshot.flatSortedItems.map(\.id) == [older.id])
   }
 
   @Test("Missing primary falls back to the newest item")
   func missingPrimaryUsesNewestItem() {
     let older = SnapshotItem(
       id: "claude-old",
-      modulePath: "/tmp/repo-a",
       timestamp: Date(timeIntervalSince1970: 100)
     )
     let newer = SnapshotItem(
       id: "codex-new",
-      modulePath: "/tmp/repo-a",
       timestamp: Date(timeIntervalSince1970: 200)
     )
 
     let snapshot = MonitoringItemsSnapshot(
       items: [older, newer],
       primaryItemID: "missing",
-      layoutMode: .single,
-      modulePath: { $0.modulePath },
       timestamp: { $0.timestamp }
     )
 
@@ -62,34 +52,29 @@ struct MonitoringItemsSnapshotTests {
     #expect(snapshot.visibleItems.map(\.id) == [newer.id])
   }
 
-  @Test("Multi-item layouts group by module and sort groups by recent activity")
-  func groupsAndSortsVisibleItems() {
+  @Test("All item IDs remain available for selection and pruning")
+  func itemIDsIncludeEveryItem() {
     let first = SnapshotItem(
       id: "first",
-      modulePath: "/tmp/repo-b",
       timestamp: Date(timeIntervalSince1970: 100)
     )
     let second = SnapshotItem(
       id: "second",
-      modulePath: "/tmp/repo-a",
       timestamp: Date(timeIntervalSince1970: 300)
     )
     let third = SnapshotItem(
       id: "third",
-      modulePath: "/tmp/repo-a",
       timestamp: Date(timeIntervalSince1970: 200)
     )
 
     let snapshot = MonitoringItemsSnapshot(
       items: [first, second, third],
       primaryItemID: nil,
-      layoutMode: .list,
-      modulePath: { $0.modulePath },
       timestamp: { $0.timestamp }
     )
 
-    #expect(snapshot.groupedItems.map(\.modulePath) == ["/tmp/repo-a", "/tmp/repo-b"])
-    #expect(snapshot.groupedItems.first?.items.map(\.id) == [second.id, third.id])
-    #expect(snapshot.flatSortedItems.map(\.id) == [second.id, third.id, first.id])
+    #expect(snapshot.effectivePrimaryItemID == second.id)
+    #expect(snapshot.visibleItems.map(\.id) == [second.id])
+    #expect(snapshot.itemIDs == [first.id, second.id, third.id])
   }
 }
